KAN-282 refactor : 상수 분리
