--- ubun2_fe/src/member/components/Order/orderDummyData.js
@@ -1,319 +1,0 @@
-export const ADDRESS_DUMMY_DATA = [
-  {
-    addressId: 1,
-    addressNickname: '우리집',
-    recipientName: '홍홍길',
-    recipientPhone: '010-2323-1212',
-    defaultStatus: true,
-    address: '경기 용인시 수지구 성복2로 158 608동 504호 (성복동, 성동마을엘지빌리지6차) 608동 504호',
-  },
-  {
-    addressId: 2,
-    addressNickname: '김막시무스네 집',
-    recipientName: '김와우',
-    recipientPhone: '010-2323-1212',
-    defaultStatus: false,
-    address: '경기 용인시 수지구 성복2로 158 608동 504호 (성복동, 성동마을엘지빌리지6차) 608동 504호',
-  },
-  {
-    addressId: 3,
-    addressNickname: '왕엘리자베스네 집',
-    recipientName: '오우예',
-    recipientPhone: '010-2323-1212',
-    defaultStatus: false,
-    address: '경기 용인시 수지구 성복2로 158 608동 504호 (성복동, 성동마을엘지빌리지6차) 608동 504호',
-  },
-];
-
-export const deliveryContent = [
-  {
-    value: 1,
-    text: '문앞',
-  },
-  {
-    value: 2,
-    text: '경비실',
-  },
-  {
-    value: 3,
-    text: '택배함',
-  },
-  {
-    value: 4,
-    text: '직접 받고 부재 시 문 앞',
-  },
-  {
-    value: 5,
-    text: '직접 받고 부재 시 경비실',
-  },
-  {
-    value: 6,
-    text: '직접 받고 부재 시 택배함',
-  },
-];
-
-export const ORDER_LIST_DUMMY_DATA = {
-  memberInfo: {
-    name: '홍길동',
-    singleOrders: 2,
-    subscriptionOrders: 1,
-    pending: 1,
-    denied: 1,
-    approved: 1,
-  },
-  orderList: [
-    {
-      orderId: 1,
-      addressNickname: null,
-      createdAt: '2024-07-17',
-      address: null,
-      orderStatus: 'PENDING',
-      paymentType: 'CARD',
-      cardName: '우리카드',
-      accountName: '우리은행',
-      cardNumber: '1111-2222-3333-4444',
-      accountNumber: '1111-2222-3333-4444',
-      orderProducts: [
-        {
-          productId: 1,
-          productName: '심심한 목장 우유 2L',
-          productDescription: '[첫구매딜] 심심한 목장 우유 2L * 2팩 무블럭?',
-          productImageOriginalName: 'milk.jpg',
-          productImagePath: 'https://thingool123.godohosting.com/data/goods/20/08/35/1000010705/1000010705_detail_023.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'PENDING',
-        },
-        {
-          productId: 2,
-          productName: '심심한 목장 우유 2L / 1개',
-          productDescription: '[첫구매딜] 심심한 목장 우유 2L * 2팩 무블럭?',
-          productImageOriginalName: 'milk.jpg',
-          productImagePath: 'https://thingool123.godohosting.com/data/goods/20/08/35/1000010705/1000010705_detail_023.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'PENDING',
-        },
-        {
-          productId: 3,
-          productName: '심심한 계란 1판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://thingool123.godohosting.com/data/goods/20/08/35/1000010705/1000010705_detail_023.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'PENDING',
-        },
-      ],
-    },
-    {
-      orderId: 2,
-      addressNickname: null,
-      createdAt: '2024-07-15',
-      address: null,
-      orderStatus: 'DENIED',
-      paymentType: 'CARD',
-      cardName: '우리카드',
-      accountName: '우리은행',
-      cardNumber: '1111-2222-3333-4444',
-      accountNumber: '1111-2222-3333-4444',
-      orderProducts: [
-        {
-          productId: 1,
-          productName: '심심한 목장 우유 2L',
-          productDescription: '[첫구매딜] 심심한 목장 우유 2L * 2팩 무블럭?',
-          productImageOriginalName: 'milk.jpg',
-          productImagePath: 'https://thingool123.godohosting.com/data/goods/20/08/35/1000010705/1000010705_detail_023.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'DENIED',
-        },
-        {
-          productId: 2,
-          productName: '심심한 목장 우유 2L / 1개',
-          productDescription: '[첫구매딜] 심심한 목장 우유 2L * 2팩 무블럭?',
-          productImageOriginalName: 'milk.jpg',
-          productImagePath: 'https://thingool123.godohosting.com/data/goods/20/08/35/1000010705/1000010705_detail_023.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'DENIED',
-        },
-        {
-          productId: 3,
-          productName: '심심한 계란 1판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://thingool123.godohosting.com/data/goods/20/08/35/1000010705/1000010705_detail_023.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'DENIED',
-        },
-      ],
-    },
-  ],
-
-  subscriptionOrderList: [
-    {
-      orderId: 3,
-      addressNickname: null,
-      createdAt: '2024-07-15',
-      address: null,
-      orderStatus: 'APPROVE',
-      paymentType: 'CARD',
-      cardName: '우리카드',
-      accountName: '우리은행',
-      cardNumber: '1111-2222-3333-4444',
-      accountNumber: '1111-2222-3333-4444',
-      orderProducts: [
-        {
-          productId: 1,
-          cycleNumber: 1,
-          productName: '심심한 계란 1판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 20000,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 18000,
-          orderProductStatus: 'APPROVE',
-          deliveryDate: '2024-07-15',
-        },
-        {
-          productId: 2,
-          cycleNumber: 1,
-          productName: '심심한 계란 1판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 20000,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 18000,
-          orderProductStatus: 'APPROVE',
-          deliveryDate: '2024-07-15',
-        },
-        {
-          productId: 3,
-          cycleNumber: 1,
-          productName: '심심한 계란 1판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 20000,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 18000,
-          orderProductStatus: 'APPROVE',
-          deliveryDate: '2024-07-15',
-        },
-        {
-          productId: 4,
-          cycleNumber: 2,
-          productName: '심심한 계란 2판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 20000,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 18000,
-          orderProductStatus: 'APPROVE',
-          deliveryDate: '2024-07-22',
-        },
-        {
-          productId: 5,
-          cycleNumber: 2,
-          productName: '심심한 계란 2판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 20000,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 18000,
-          orderProductStatus: 'APPROVE',
-          deliveryDate: '2024-07-22',
-        },
-        {
-          productId: 6,
-          cycleNumber: 2,
-          productName: '심심한 계란 2판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 20000,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 18000,
-          orderProductStatus: 'APPROVE',
-          deliveryDate: '2024-07-22',
-        },
-      ],
-    },
-    {
-      orderId: 4,
-      addressNickname: null,
-      createdAt: '2024-07-20',
-      address: null,
-      orderStatus: 'APPROVE',
-      paymentType: 'CARD',
-      cardName: '우리카드',
-      accountName: '우리은행',
-      cardNumber: '1111-2222-3333-4444',
-      accountNumber: '1111-2222-3333-4444',
-      orderProducts: [
-        {
-          productId: 1,
-          cycleNumber: 1,
-          productName: '다른 계란 2판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'APPROVE',
-        },
-        {
-          productId: 2,
-          cycleNumber: 1,
-          productName: '다른 계란 2판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'APPROVE',
-        },
-        {
-          productId: 3,
-          productName: '다른 계란 2판',
-          productDescription: '[첫구매딜] 심심한 계란 1판 * 3팩 무블럭?',
-          productImageOriginalName: 'egg.jpg',
-          productImagePath: 'https://forestlife.shop/data/item/1622583258/2342CD34556020C510.jpg',
-          productPrice: 19900,
-          quantity: 1,
-          productDiscount: 10,
-          totalPrice: 19900,
-          orderProductStatus: 'APPROVE',
-        },
-      ],
-    },
-  ],
-};
