+++ ubun2_fe/src/member/components/Order/orderModalContent.js
@@ -1,0 +1,33 @@
+export const deliveryContent = [
+  {
+    value: 1,
+    text: '문앞',
+  },
+  {
+    value: 2,
+    text: '경비실',
+  },
+  {
+    value: 3,
+    text: '택배함',
+  },
+  {
+    value: 4,
+    text: '직접 받고 부재 시 문 앞',
+  },
+  {
+    value: 5,
+    text: '직접 받고 부재 시 경비실',
+  },
+  {
+    value: 6,
+    text: '직접 받고 부재 시 택배함',
+  },
+];
+
+export const myOrderPeriodOptions = [
+  { value: null, label: '전체', periodType: null, periodValue: null },
+  { value: 'WEEK', label: '1주일', periodType: 'WEEK', periodValue: 1 },
+  { value: 'MONTH', label: '1달', periodType: 'MONTH', periodValue: 1 },
+  { value: 'THREE_MONTHS', label: '3달', periodType: 'MONTH', periodValue: 3 },
+];
